requirements.txt
aniso8601==9.0.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask[async]
Flask-RESTful==0.3.10
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyJWT==2.8.0
pymssql==2.2.11
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
